# cmd/docs/swagger.yaml
basePath: /
definitions:
...
  models.SetPasswordRequest:
    properties:
      password:
        description: New password
        example: new_password
        type: string
    required:
    - password
    type: object
...
  models.UserRespModel:
    description: User model containing all user-related information.
    properties:
      created_time:
        description: Time when the user was created
        example: "2023-04-12T15:04:05Z"
        type: string
      mobile_number:
        description: Mobile number of the user
        example: "9876543210"
        type: string
      org_status:
        allOf:
        - $ref: '#/definitions/models.OrganisationStatus'
        description: Organization ID
        example: "123456"
      org_uuid:
        description: UUID of the organization
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      remarks:
        description: Additional remarks about the user
        example: User is active and verified
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Role of the user
        example: Admin
      status:
        allOf:
        - $ref: '#/definitions/models.UserStatus'
        description: Status of the user
        example: Active
      uid:
        description: Auto-incremented unique identifier
        example: 123e4567-e89b-12d3-a456-426614174111
        type: string
      update_history:
        description: History of updates
        items:
          $ref: '#/definitions/models.UpdateHistory'
        type: array
      updated_time:
        description: Time when the user was last updated
        example: "2023-04-12T15:04:05Z"
        type: string
      userid:
        description: Unique identifier for the user
        example: "112345"
        type: string
      username:
        description: Username of the user
        example: john_doe
        type: string
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: This is the API documentation for the Fverify backend.
  title: FVerify API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Retrieve all users in the system
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organisation Id
        in: header
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserRespModel'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.InvalidAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user in the system
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organisation Id
        in: header
        name: org_id
        required: true
        type: string
      - description: User data (all fields are mandatory)
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserReqModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserRespModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.InvalidAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Create a new user
      tags:
      - Users
  /api/v1/users/{userId}:
    get:
      consumes:
      - application/json
      description: Retrieve a user by their unique ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organisation Id
        in: header
        name: org_id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserRespModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.InvalidAuthResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Get a user by userId
      tags:
      - Users
  /api/v1/users/admin/create:
    post:
      consumes:
      - application/json
      description: Create a new admin user in the system (requires API key)
      parameters:
      - description: API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Admin user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserReqModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserRespModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Create a new admin user
      tags:
      - Users
  /api/v1/users/login:
    post:
      consumes:
      - application/json
      description: Validate username and password, and return user details with a
        token
      parameters:
      - description: Login credentials
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.InvalidAuthResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Login a user
      tags:
      - Users
  /api/v1/users/owner/create:
    post:
      consumes:
      - application/json
      description: Create a new admin user in the system (requires API key)
      parameters:
      - description: API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: User data (all fields are mandatory)
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserReqModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserRespModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Create a new owner
      tags:
      - Users
  /api/v1/users/roles:
    get:
      consumes:
      - application/json
      description: Retrieve all user roles for a given organisation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.InvalidAuthResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Get user roles
      tags:
      - Users
  /api/v1/users/uid/{uId}:
    put:
      consumes:
      - application/json
      description: Update an existing user's details
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organisation Id
        in: header
        name: org_id
        required: true
        type: string
      - description: User uId
        in: path
        name: uId
        required: true
        type: string
      - description: User data (all fields are mandatory)
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserReqModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserRespModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.InvalidAuthResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Update a user
      tags:
      - Users
  /api/v1/users/uid/{uId}/setpassword:
    put:
      consumes:
      - application/json
      description: Set a new password for a user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organisation Id
        in: header
        name: org_id
        required: true
        type: string
      - description: User uId
        in: path
        name: uId
        required: true
        type: integer
      - description: New password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.SetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.InvalidAuthResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Set a user's password
      tags:
      - Users
  /api/v1/users/userid/{userId}:
    delete:
      description: Delete a user by their unique userId
      parameters:
      - description: User userId
        in: path
        name: userId
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Organisation Id
        in: header
        name: org_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.InvalidAuthResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.InternalErrorResponse'
      summary: Delete a user by userId
      tags:
      - Users
schemes:
- http
swagger: "2.0"